addressed all kurt and junaid issues
