+++ build.gradle
@@ -105,7 +105,7 @@ pmd {
 }
 
 jacoco {
-  toolVersion = "0.8.6"
+  toolVersion = "0.8.8"
 }
 
 jacocoTestReport {
