+++ src/integrationTest/java/uk/gov/hmcts/reform/channel/management/config/SwaggerPublisherTest.java
@@ -1,0 +1,45 @@
+package uk.gov.hmcts.reform.channel.management.config;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+import uk.gov.hmcts.reform.channel.management.Application;
+
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+/**
+ * Built-in feature which saves service's swagger specs in temporary directory.
+ * Each travis run on master should automatically save and upload (if updated) documentation.
+ */
+@SpringBootTest(classes = {Application.class},
+    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+class SwaggerPublisherTest {
+
+    @Autowired
+    private MockMvc mvc;
+
+    @DisplayName("Generate swagger documentation")
+    @Test
+    @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
+    void generateDocs() throws Exception {
+        byte[] specs = mvc.perform(get("/v3/api-docs"))
+            .andExpect(status().isOk())
+            .andReturn()
+            .getResponse()
+            .getContentAsByteArray();
+
+        try (OutputStream outputStream = Files.newOutputStream(Paths.get("/tmp/swagger-specs.json"))) {
+            outputStream.write(specs);
+        }
+
+    }
+}
