+++ src/main/java/uk/gov/hmcts/reform/channel/management/config/SpringSecurityConfig.java
@@ -1,0 +1,23 @@
+package uk.gov.hmcts.reform.channel.management.config;
+
+import com.azure.spring.cloud.autoconfigure.aad.AadResourceServerWebSecurityConfigurerAdapter;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+
+@EnableWebSecurity
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+@Profile("!dev")
+@SuppressWarnings("java:S4502")
+public class SpringSecurityConfig extends AadResourceServerWebSecurityConfigurerAdapter {
+
+    /**
+     * Add configuration logic as needed.
+     */
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.csrf().disable();
+        super.configure(http);
+    }
+}
