+++ src/main/java/uk/gov/hmcts/reform/channel/management/controllers/RootController.java
@@ -1,0 +1,30 @@
+package uk.gov.hmcts.reform.channel.management.controllers;
+
+import io.swagger.annotations.Api;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import static org.springframework.http.ResponseEntity.ok;
+
+/**
+ * Default endpoints per application.
+ */
+@RestController
+@Api(tags = "Channel Management root API")
+public class RootController {
+
+    /**
+     * Root GET endpoint.
+     *
+     * <p>Azure application service has a hidden feature of making requests to root endpoint when
+     * "Always On" is turned on.
+     * This is the endpoint to deal with that and therefore silence the unnecessary 404s as a response code.
+     *
+     * @return Welcome message from the service.
+     */
+    @GetMapping("/")
+    public ResponseEntity<String> welcome() {
+        return ok("Welcome to pip-channel-management");
+    }
+}
