+++ src/main/java/uk/gov/hmcts/reform/channel/management/models/external/subscriptionmanagement/SearchType.java
@@ -1,0 +1,10 @@
+package uk.gov.hmcts.reform.channel.management.models.external.subscriptionmanagement;
+
+public enum SearchType {
+    /**
+     * This determines the type of search or addition we are using.
+     */
+    COURT_ID,
+    CASE_ID,
+    CASE_URN
+}
