--- src/main/java/uk/gov/hmcts/reform/demo/models/Channel.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.demo.models;
-
-public enum Channel {
-    // add as required
-    EMAIL,
-    API;
-}
