--- src/main/java/uk/gov/hmcts/reform/demo/models/SearchType.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.reform.demo.models;
-
-public enum SearchType {
-    /**
-     * This determines the type of search or addition we are using.
-     */
-    COURT_ID,
-    CASE_ID,
-    CASE_URN
-}
