--- src/main/java/uk/gov/hmcts/reform/demo/services/SubscriptionManagementService.java
@@ -1,28 +1,0 @@
-package uk.gov.hmcts.reform.demo.services;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.demo.models.Subscription;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-@Slf4j
-@Component
-public class SubscriptionManagementService {
-
-    public Map<String, List<Subscription>> deduplicateSubscriptions(List<Subscription> listOfSubs) {
-        Map<String, List<Subscription>> mapOfSubscriptions = new ConcurrentHashMap<>();
-        listOfSubs.forEach(subscription -> {
-            List<Subscription> currentList = new ArrayList<>();
-            if (mapOfSubscriptions.get(subscription.getUserId()) != null) {
-                currentList = mapOfSubscriptions.get(subscription.getUserId());
-            }
-            currentList.add(subscription);
-            mapOfSubscriptions.put(subscription.getUserId(), currentList);
-        });
-        return mapOfSubscriptions;
-    }
-}
