+++ src/test/java/uk/gov/hmcts/reform/channel/management/config/WebClientConfigurationTest.java
@@ -1,0 +1,17 @@
+package uk.gov.hmcts.reform.channel.management.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+@Profile("test")
+public class WebClientConfigurationTest {
+
+    @Bean
+    public WebClient webClient() {
+        return WebClient.create();
+    }
+
+}
