+++ src/test/java/uk/gov/hmcts/reform/channel/management/errorhandling/GlobalExceptionHandlerTest.java
@@ -1,0 +1,35 @@
+package uk.gov.hmcts.reform.channel.management.errorhandling;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.mock.web.MockHttpServletRequest;
+import org.springframework.web.context.request.ServletWebRequest;
+import uk.gov.hmcts.reform.channel.management.errorhandling.exceptions.ChannelNotFoundException;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+class GlobalExceptionHandlerTest {
+
+    @Test
+    void testHandleChannelNotFoundMethod() {
+
+        GlobalExceptionHandler globalExceptionHandler = new GlobalExceptionHandler();
+
+        ChannelNotFoundException channelNotFoundException
+            = new ChannelNotFoundException("This is a test message");
+
+
+        MockHttpServletRequest mockHttpServletRequest = new MockHttpServletRequest();
+        ServletWebRequest servletWebRequest = new ServletWebRequest(mockHttpServletRequest);
+
+        ResponseEntity<ExceptionResponse> responseEntity =
+            globalExceptionHandler.handleChannelNotFound(channelNotFoundException, servletWebRequest);
+
+        assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode(), "Status code should be not found");
+        assertNotNull(responseEntity.getBody(), "Response should contain a body");
+        assertEquals("This is a test message", responseEntity.getBody().getMessage(),
+                     "The message should match the message passed in");
+    }
+}
