--- src/test/java/uk/gov/hmcts/reform/rsecheck/config/WebClientConfigurationTest.java
@@ -1,17 +1,0 @@
-package uk.gov.hmcts.reform.rsecheck.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-@Profile("test")
-public class WebClientConfigurationTest {
-
-    @Bean
-    public WebClient webClient() {
-        return WebClient.create();
-    }
-
-}
