--- src/test/java/uk/gov/hmcts/reform/rsecheck/config/WebClientCreationTest.java
@@ -1,42 +1,0 @@
-package uk.gov.hmcts.reform.rsecheck.config;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository;
-import org.springframework.web.reactive.function.client.WebClient;
-import uk.gov.hmcts.reform.demo.config.WebClientConfig;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-@ExtendWith(MockitoExtension.class)
-class WebClientCreationTest {
-
-    @Mock
-    ClientRegistrationRepository clientRegistrationRepository;
-
-    @Mock
-    OAuth2AuthorizedClientRepository clientRepository;
-
-    @Test
-    void createWebClient() {
-
-        WebClientConfig webClientConfiguration = new WebClientConfig();
-        WebClient webClient =
-            webClientConfiguration.webClient(clientRegistrationRepository, clientRepository);
-
-        assertNotNull(webClient, "WebClient has not been created successfully");
-    }
-
-    @Test
-    void createWebClientInsecure() {
-
-        WebClientConfig webClientConfiguration = new WebClientConfig();
-        WebClient webClient =
-            webClientConfiguration.webClientInsecure();
-
-        assertNotNull(webClient, "WebClient has not been created successfully");
-    }
-}
